gateway_server/requirements.txt
fastapi
uvicorn[standard]
grpcio
protobuf
jinja2
